Make PreserveDependency deprecated
We will keep supporting it for now, and remove support for it once
runtime has been converted to DynamicDependency
