+++ src/linker/Linker/DependencyInfo.cs
@@ -110,9 +110,7 @@ namespace Mono.Linker
 
 		// Linker-specific behavior (preservation hints, patterns, user inputs, linker outputs, etc.)
 		DynamicDependency = 67, // DynamicDependency attribute -> member
-#if !FEATURE_ILLINK
 		PreservedDependency = 67, // PreserveDependency attribute -> member
-#endif
 		AccessedViaReflection = 68, // method -> detected member accessed via reflection from that method
 		PreservedMethod = 69, // type/method -> preserved method (explicitly preserved in Annotations by XML or other steps)
 		TypePreserve = 70, // type -> field/method preserved for the type (explicitly set in Annotations by XML or other steps)
